[Feature] add integration to support skytether (#3)
* task: updated buf config

This updates syntax to accommodate new version of buf, as well as
updates the version of protobuf we depend on

* task: add specification for service topology

These additions are useful so that mohair query services *can*
accommmodate topology information in query routing and optimization. The
only necessary aspect of the topology specification is to be able add
services to a CS system which can be done independently of the algebra
portion of the mohair specification.

* task: changed `MohairService` to `ServiceConfig`

`MohairService` is already the name of the base class for mohair
services, so the name change is for convenience

* task: added `is_active` attribute

This is so that a topology definition can be fixed but services within
the topology can be activated/de-activated (sparsely running)

* task: updated buf gen config

Updated to buf v2 and updated protobuf version to 27.3

* task: moved proto files to new directory

this is to reflect changes to package name

* task: changed package names and updated includes

added skytether prefix to package name and refactored includes of
substrait proto definitions to accommodate the new package name.

Also, changing dependency to buf.build/octalene registry

* task: updated protobuf version (v27.3 -> v28.2)

* chore(buf): updated config and deps

* chore(buf): updated buf configuration

* fix(buf): fixed buf configuration

* task: added Rel types for partitions and slices

This is to accommodate direct naming in substrait translation. `SkyRel`
is still usable for catalog lookups if using skytether extension

* task: updated attributes of skytether rel types

I commented that `SkyRel` would be good for catalog lookups, but the
attributes across SkyRel, SkyPartitionRel, and SkySliceRel didn't
reflect that idea. This change updates the rel type messages so that:
- SkyPartitionRel specifies a logical partition that should be resolved
  to many slices
- SkySliceRel specifies a partition slice that can be read directly.
  This may also be for a partition stored directly as a data slice
  (instead of a meta slice and a data slice)
- SkyRel specifies a partition name that should be resolved via the
  catalog (only a skytether storage extension will be able to resolve
  the lookup if data is stored in a skytether system)

* minor: moved SkyRel above other Sky* rel types

This is because `SkyRel` should still be the most generic representation
of a table or relation in skytether. `SkyPartitionRel` and `SkySliceRel`
represent a more opinionated specification of where a table or relation
can be found and how it is stored.