task: added new result operators
The SkyResultRel is designed for use in a pushback plan, describing how
the query results can be pulled (for now, just contains a context ID
corresponding to the flight "ticket").

The SkyLakeRel is designed for future use in both pushdown and pushback
plans. In a pushdown plan, it specifies information for pushing data. In
a pushback plan, it is used to know where to pull local results from
(local to the upstream service because the downstream service should
have pushed results there).

Other changes were made to simplify some of the comments and make the
overall organization of messages a bit easier to understand.