task: add fields to SkyResultRel
This operator hold information for a query context (context_id and
location) as well as information for the schema and common information
for the result. The purpose is so that a pushback plan can be translated
and executed before the result is available at the upstream service. The
schema must be known in the pushback otherwise an execution plan can't
be constructed and the superplan will be missing information to
correctly select inputs.